Moved files after merge conflict. Added Routers around test components.
