+++ WhaleSpotting/react-app/src/tests/components/Home.test.tsx
@@ -1,16 +1,17 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import Home from "../components/Home";
+import Home from "../../components/Home";
+import { BrowserRouter as Router } from "react-router-dom";
 
 describe("Home Page tests", () => {
     test("Renders Recent Sightings", () => {
-        render(<Home />);
+        render(<Router><Home /></Router>);
         const text = screen.getByText("Recent Sightings");
         expect(text).toBeInTheDocument();
     });
 
     test("Renders Filter button", () => {
-        render(<Home />);
+        render(<Router><Home /></Router>);
         const filter = screen.getByTestId("filter-button");
         expect(filter).toBeInTheDocument();
     });
